+++ frontend/components/workflows/image-generation/select.tsx
