+++ frontend/config.json
@@ -1,3 +1,3 @@
 {
-  "api_base_url": "https://o14iiwf5ime2n2-64410ae5-80.proxy.runpod.io"
+  "api_base_url": "https://cjg40wkwbr6ogj-80.proxy.runpod.net"
 }
